% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga}
\alias{blupga}
\title{Standard GBLUP}
\usage{
blupga(G, S, phenodata, valset, verbose = T)
}
\arguments{
\item{G}{GRM constructed using all available SNPs and all samples Defaults to NULL. Use \code{make_GRM()} to get this.}

\item{S}{Weighted G-matrix constructed using only selected SNPs and all samples Defaults to NULL. Use \code{make_weighted_GRM()} to get this.}

\item{phenodata}{data frame with 2 columns. One col must be named 'ID' and contain sample IDs. Another column must be named 'y' and contain the phenotypes. Defaults to NULL.}

\item{valset}{vector of indices that defines which rows in \code{phenodata} will be set to NA and used for cross-validation Defaults to NULL.}
}
\value{
a data frame containing the correlation between the predicted phenotype and the true phenotype of the individuals in the \code{valset}. Each row contains that correlation at increasing values of the weighting value W.
}
\description{
This function runs the BLUP|GA method where certain SNPs are weighted in the a special GRM (S) prior to GBLUP.
}
\examples{
blupga()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
