% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga_CAND}
\alias{blupga_CAND}
\title{BLUP|GA using SNPs in Candidate Genes}
\usage{
blupga_CAND(G = NULL, phenodata = NULL, valset = NULL, GRMs = NULL,
  verbose = T)
}
\arguments{
\item{G}{G-matrix constructed using all available SNPs and samples Defaults to NULL.}

\item{phenodata}{data frame with 2 columns: ID and y Defaults to NULL.}

\item{valset}{vector of indices Defaults to NULL.}

\item{GRMs}{List of G-matrices, each constructed from just the SNPs in one candidate gene/region Defaults to NULL.}
}
\description{
This function runs the BLUP|GA method where SNPs in candidate genes are weighted in the GRM prior to GBLUP
}
\examples{
blupga_CAND()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
