% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga_EFF}
\alias{blupga_EFF}
\title{BLUP|GA using top SNPS according to estimated SNP effect}
\usage{
blupga_EFF(G, phenodata, valset = NULL, genomat, bsq, perc, flank = TRUE)
}
\arguments{
\item{G}{G-matrix constructed using all available SNPs and samples Defaults to NULL.}

\item{phenodata}{data frame with 2 columns: ID and y Defaults to NULL.}

\item{valset}{vector of indices Defaults to NULL.}

\item{perc}{proportion of SNPs to be weighted (between 0 and 1), where 0.05 means the top 5% of SNPs will be weighted. Defaults to NULL.}
}
\description{
This function runs the BLUP|GA method where SNPs with the greatest squared effect size are weighted in the GRM prior to GBLUP
}
\examples{
blupga_EFF()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
