% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{est_SNPeffects}
\alias{est_SNPeffects}
\title{run a GWAS using a training set of samples to estimate marker effects}
\usage{
est_SNPeffects(phenodata, genomat, valset, fixmat = NULL)
}
\arguments{
\item{phenodata}{data frame with 2 columns: ID and y Defaults to NULL.}

\item{genomat}{matrix of genotypes in -1,0,1 format (i.e. 0 is heterozygous, 1 and -1 are opposing homozygous states)}

\item{valset}{vector of indices Defaults to NULL.}
}
\description{
This function runs the BLUP|GA method where SNPs with the greatest squared effect size are weighted in the GRM prior to GBLUP
}
\examples{
est_SNPeffects()
}
