% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M}
\alias{M}
\title{example genotype matrix in {-1,0,1} format}
\format{An object of class \code{matrix} with 100 rows and 500 columns.}
\usage{
M
}
\description{
A dataset with an example raw genotype matrix for 100 individuals and 500 SNPs
Dimensions [100,500]
}
\keyword{datasets}
