% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{make_GRM}
\alias{make_GRM}
\title{Construct the G matrix from genotypes}
\usage{
make_GRM(genomat)
}
\arguments{
\item{genomat}{matrix of genotypes in -1,0,1 format (i.e. 0 is heterozygous, 1 and -1 are opposing homozygous states)}
}
\description{
This function runs the BLUP|GA method where SNPs with the greatest squared effect size are weighted in the GRM prior to GBLUP
}
\examples{
make_GRM()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
