% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{example phenotype data}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
pheno
}
\description{
A pretend dataset with one phenotype measured in 100 individuals
\describe{
  \item{ID}{individual ID}
  \item{y}{measured phenotype}
  ...
}
}
\keyword{datasets}
