% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga}
\alias{blupga}
\title{Standard GBLUP}
\usage{
blupga(G, S, phenodata, valset, verbose = T)
}
\arguments{
\item{G}{G-matrix constructed using all available SNPs and all samples Defaults to NULL.}

\item{S}{Weighted G-matrix constructed using only selected SNPs and all samples Defaults to NULL.}

\item{phenodata}{data frame with 2 columns: ID and y Defaults to NULL.}

\item{valset}{vector of indices Defaults to NULL.}
}
\description{
This function runs the BLUP|GA method where SNPs in candidate genes are weighted in the GRM prior to GBLUP
}
\examples{
blupga()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
