% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga_CAND}
\alias{blupga_CAND}
\title{BLUP|GA using SNPs in Candidate Genes}
\usage{
blupga_CAND(G, phenodata, valset, GRMs = NULL, verbose = T)
}
\arguments{
\item{G}{G-matrix constructed using all available SNPs and samples Defaults to NULL.}

\item{valset}{vector of indices that defines which rows in \code{phenodata} will be set to NA and used for cross-validation Defaults to NULL.}

\item{GRMs}{List of G-matrices, each constructed from just the SNPs in one candidate gene/region. Defaults to NULL.}

\item{@param}{phenodata data frame with 2 or 3 columns. One col must be named 'ID' and contain sample IDs. Another col must be named 'y' and contain the phenotypes. If fixed effects are included then a 3rd col called 'FE' should contain the categorical effects. Defaults to NULL.}
}
\value{
A data frame containing the correlation between the predicted phenotype and the true phenotype of the individuals in the \code{valset}.

 Since BLUP|GA is run for each value of omega (W) from 0.0 to 1.0 in increments of 0.10, each row of the returned data frame
  contains the cross-validation correlation at one value of omega (W). This allows the user to find the value of W
  at which the predictive ability (COR) is maximised.
\describe{
  \item{W}{omega weighting for selected SNPS in candidate genes (0.0--1.0)}
  \item{COR}{cross validation predictive ability (0.0--1.0)}
  }
}
\description{
This function runs the BLUP|GA method where SNPs in candidate genes are weighted in the GRM prior to GBLUP
}
\examples{
blupga_CAND()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
