% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga_GBLUP}
\alias{blupga_GBLUP}
\title{Standard GBLUP}
\usage{
blupga_GBLUP(G, phenodata, valset, verbose = T)
}
\arguments{
\item{G}{GRM constructed using all available SNPs and all samples Defaults to NULL. Use \code{make_GRM()} to get this.}

\item{phenodata}{data frame with 2 columns. One col must be named 'ID' and contain sample IDs. Another column must be named 'y' and contain the phenotypes. Defaults to NULL.}

\item{valset}{vector of indices that defines which rows in \code{phenodata} will be set to NA and used for cross-validation Defaults to NULL.}
}
\description{
This function runs straightforward GBLUP using the rrBLUP package. This is the equivalent of BLUP|GA where W=0 (i.e. no SNPs are weighted)
}
\examples{
blupga_GBLUP()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
