% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{blupga_GBLUP}
\alias{blupga_GBLUP}
\title{Standard GBLUP}
\usage{
blupga_GBLUP(G, phenodata, valset, verbose = T)
}
\arguments{
\item{G}{GRM constructed using all available SNPs and all samples Defaults to NULL. Use \code{make_GRM()} to get this.}

\item{phenodata}{data frame with 2 or 3 columns. One col must be named 'ID' and contain sample IDs. Another col must be named 'y' and contain the phenotypes. If fixed effects are included then a 3rd col called 'FE' should contain the categorical effects. Defaults to NULL.}

\item{valset}{vector of indices that defines which rows in \code{phenodata} will be set to NA and used for cross-validation Defaults to NULL.}
}
\value{
a data frame containing the correlation between the predicted phenotype and the true phenotype of the individuals in the valset.
\describe{
  \item{W}{omega weighting for selected SNPS in candidate genes (0.0--1.0)}
  \item{COR}{cross validation predictive ability (0.0--1.0)}
  }
}
\description{
This function runs straightforward GBLUP using the rrBLUP package. This is the equivalent of BLUP|GA where W=0 (i.e. no SNPs are weighted)
}
\examples{
# get example genotype matrix and phenotype data
data(M)
data(pheno)

G <- make_GRM(M)

# choose a validation set of 20 random individuals
val <- sample(1:nrow(pheno), 20)
results <- blupga_GBLUP(G, pheno, val)
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
