% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{est_SNPeffects}
\alias{est_SNPeffects}
\title{estimate the additive effect of each SNP upon a trait (i.e. quick GWAS)}
\usage{
est_SNPeffects(phenodata, genomat, valset, fixmat = NULL)
}
\arguments{
\item{phenodata}{data frame with 2 columns. One col must be named 'ID' and contain sample IDs. Another column must be named 'y' and contain the phenotypes. Defaults to NULL.}

\item{genomat}{matrix of genotypes in -1,0,1 format (i.e. 0 is heterozygous, 1 and -1 are opposing homozygous states)}

\item{valset}{vector of indices Defaults to NULL.}
}
\value{
a vector of squared SNP effects, one per SNP in \code{genomat}, estimated with the EMMAX GWAS method from the \pkg{cpgen} package.
}
\description{
estimate the additive effect of each SNP upon a trait (i.e. quick GWAS). SNPs are provided in a genotype matrix.
}
\examples{
# get an example genotype matrix
data(M)
# get an example phenotype data frame
data(pheno)
# choose a validation set of 20 random individuals
val <- sample(1:nrow(pheno), 20)
est_SNPeffects(pheno, M, val)
}
