% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_weighted_GRM}
\alias{make_weighted_GRM}
\title{Construct a weighted G matrix from genotypes}
\usage{
make_weighted_GRM(genomat, weights)
}
\arguments{
\item{genomat}{matrix of genotypes in -1,0,1 format (i.e. 0 is heterozygous, 1 and -1 are opposing homozygous states)}

\item{weights}{vector of weights for each of the SNPs in the \code{genomat}. Use a vector of 1s to avoid weighting.}
}
\value{
a weighted Genomic Relationship Matrix with cols and rows named according to the rownames in \code{genomat}. Uses the \pkg{cpgen} package implementation of Van Radens GRM method
}
\description{
This function ...
}
\examples{
# get an example genotype matrix
data(M)
# generate weights for the SNPs
wt <- runif(ncol(M), min = 1, max = 10)

make_weighted_GRM(M, wt)
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
