% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPGA.R
\name{make_weighted_GRM}
\alias{make_weighted_GRM}
\title{Construct a weighted G matrix from genotypes}
\usage{
make_weighted_GRM(genomat, weights)
}
\arguments{
\item{genomat}{matrix of genotypes in -1,0,1 format (i.e. 0 is heterozygous, 1 and -1 are opposing homozygous states)}

\item{weights}{vector of weights for each of the SNPs in the geno matrix}
}
\description{
This function ...
}
\examples{
make_weighted_GRM()
}
\keyword{GBLUP,BLUP|GA,SNP}
\keyword{selection}
